requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
qdrant-client
openai
python-dotenv
pydantic-settings
python-multipart
pypdf
langchain
langchain-openai
celery
redis
langchain-text-splitters
groq
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]